backend/requirements.txt
fastapi
uvicorn
boto3
python-multipart
pydantic
python-dotenv
pandas
